Add notice of changes to changed file
The terms of the ECL-2.0 license under which EKV 302 is distributed
require that any modifications of the files be prominently documented
in a notice.  I am adding just such a notice.

I will note, however, that the repository from which I forked this
repo and on which I base my changes has itself been modified from the
original, and does not (as far as I can tell) carry a similar notice
of changes as required by the license.  Should that be added in the
future, I will rebase my changes onto that fork's update.